requirements.txt
pypdf    # pdf to text
langchain_text_splitters
langchain-community
openai
huggingface_hub
python-dotenv
pandas
beautifulsoup4
lxml

